Tests: move plugin dir to plugins dir
We should be consistent in our naming for classes and resources.
