+++ src/test/java/org/elasticsearch/plugins/PluginLuceneCheckerTests.java
@@ -17,12 +17,11 @@
  * under the License.
  */
 
-package org.elasticsearch.plugin;
+package org.elasticsearch.plugins;
 
 import com.google.common.collect.Lists;
 import org.elasticsearch.action.admin.cluster.node.info.NodesInfoResponse;
 import org.elasticsearch.nodesinfo.SimpleNodesInfoTests;
-import org.elasticsearch.plugins.PluginsService;
 import org.elasticsearch.test.ElasticsearchIntegrationTest;
 import org.elasticsearch.test.ElasticsearchIntegrationTest.ClusterScope;
 import org.elasticsearch.test.hamcrest.ElasticsearchAssertions;
@@ -49,7 +48,7 @@ public class PluginLuceneCheckerTests extends ElasticsearchIntegrationTest {
                 settingsBuilder().put("plugins.check_lucene", false)
                         .put("plugins." + PluginsService.ES_PLUGIN_PROPERTIES_FILE_KEY, "es-plugin-test.properties")
                         .put("plugins." + PluginsService.LOAD_PLUGIN_FROM_CLASSPATH, true).build(),
-                "/org/elasticsearch/plugin/lucene/");
+                "/org/elasticsearch/plugins/lucene/");
         logger.info("--> server {} started" + serverNodeId);
 
         NodesInfoResponse response = client().admin().cluster().prepareNodesInfo().clear().setPlugins(true).execute().actionGet();
@@ -74,7 +73,7 @@ public class PluginLuceneCheckerTests extends ElasticsearchIntegrationTest {
                 settingsBuilder().put("plugins.check_lucene", true)
                         .put("plugins." + PluginsService.ES_PLUGIN_PROPERTIES_FILE_KEY, "es-plugin-test.properties")
                         .put("plugins." + PluginsService.LOAD_PLUGIN_FROM_CLASSPATH, true).build(),
-                "/org/elasticsearch/plugin/lucene/");
+                "/org/elasticsearch/plugins/lucene/");
         logger.info("--> server {} started" + serverNodeId);
 
         NodesInfoResponse response = client().admin().cluster().prepareNodesInfo().clear().setPlugins(true).execute().actionGet();
