+++ src/test/java/org/elasticsearch/plugins/ResponseHeaderPluginTests.java
@@ -16,18 +16,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.plugin;
+package org.elasticsearch.plugins;
 
 import org.apache.http.impl.client.HttpClients;
 import org.elasticsearch.common.settings.ImmutableSettings;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.http.HttpServerTransport;
-import org.elasticsearch.plugin.responseheader.TestResponseHeaderPlugin;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.test.ElasticsearchIntegrationTest;
 import org.elasticsearch.test.ElasticsearchIntegrationTest.ClusterScope;
 import org.elasticsearch.test.rest.client.http.HttpRequestBuilder;
 import org.elasticsearch.test.rest.client.http.HttpResponse;
+import org.elasticsearch.plugins.responseheader.TestResponseHeaderPlugin;
 import org.junit.Test;
 
 import static org.elasticsearch.test.ElasticsearchIntegrationTest.Scope;
