+++ src/test/java/org/elasticsearch/plugins/lucene/current/es-plugin-test.properties
@@ -16,6 +16,6 @@
 # specific language governing permissions and limitations
 # under the License.
 ################################################################
-plugin=org.elasticsearch.plugin.lucene.current.CurrentLucenePlugin
+plugin=org.elasticsearch.plugins.lucene.current.CurrentLucenePlugin
 version=2.0.0
 lucene=${lucene.version}
