+++ src/test/java/org/elasticsearch/plugins/lucene/newer/es-plugin-test.properties
@@ -16,6 +16,6 @@
 # specific language governing permissions and limitations
 # under the License.
 ################################################################
-plugin=org.elasticsearch.plugin.lucene.newer.NewerLucenePlugin
+plugin=org.elasticsearch.plugins.lucene.newer.NewerLucenePlugin
 version=3.0.0
 lucene=99.0.0
