+++ src/test/java/org/elasticsearch/plugins/lucene/old/es-plugin-test.properties
@@ -16,6 +16,6 @@
 # specific language governing permissions and limitations
 # under the License.
 ################################################################
-plugin=org.elasticsearch.plugin.lucene.old.OldLucenePlugin
+plugin=org.elasticsearch.plugins.lucene.old.OldLucenePlugin
 version=1.0.0
 lucene=3.0.0
