+++ src/test/java/org/elasticsearch/plugins/responseheader/TestResponseHeaderRestAction.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.plugin.responseheader;
+package org.elasticsearch.plugins.responseheader;
 
 import org.elasticsearch.client.Client;
 import org.elasticsearch.common.inject.Inject;
