+++ src/test/resources/org/elasticsearch/plugins/plugin_folder_file.zip
