+++ src/test/resources/org/elasticsearch/plugins/plugin_folder_site.zip
