+++ src/test/resources/org/elasticsearch/plugins/plugin_single_folder.zip
