+++ src/test/resources/org/elasticsearch/plugins/plugin_with_bin_and_config.zip
