+++ src/test/resources/org/elasticsearch/plugins/plugin_with_sourcefiles.zip
