+++ src/test/resources/org/elasticsearch/plugins/plugin_without_folders.zip
